Remove HighlightWrap theme key, restore Wrap prop on TextHighlight
