--- src/layout/RemoveTagIcon.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-let RemoveTagIcon = props => (
-  <span className="tags-input-tag-remove" {...props}>
-    x
-  </span>
-)
-
-export default RemoveTagIcon
