--- src/themes/blueberry/HighlightWrap.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-let HighlightWrap = ({ style = {}, ...props }) => (
-  <b style={{ backgroundColor: 'yellow', ...style }} {...props} />
-)
-
-export default HighlightWrap
