+++ src/themes/blueberry/index.js
@@ -10,8 +10,8 @@ import Table from './Table'
 import Button from './Button'
 import ButtonRadio from './ButtonRadio'
 import ListItem from './ListItem'
-import HighlightWrap from './HighlightWrap'
-import TagComponent from './TagComponent'
+import TextHighlight from './TextHighlight'
+import Tag from './Tag'
 import PagerItem from './PagerItem'
 
 export let Pager = props => (
@@ -22,7 +22,7 @@ let baseTheme = {
   Button,
   Checkbox,
   Fonts,
-  HighlightWrap,
+  TextHighlight,
   Input,
   ListItem,
   Pager,
@@ -30,7 +30,7 @@ let baseTheme = {
   RadioList: ButtonRadio,
   Style,
   Table,
-  TagComponent,
+  Tag,
 }
 
 let { TypeMap, ...ExampleTypes } = ExampleTypeConstructor(baseTheme)
@@ -41,13 +41,13 @@ export {
   Checkbox,
   ExampleTypes,
   Fonts,
-  HighlightWrap,
+  TextHighlight,
   Input,
   ListItem,
   PagerItem,
   Style,
   Table,
-  TagComponent,
+  Tag,
   TypeMap,
 }
