--- src/themes/greyVest/RemoveTagIcon.js
@@ -1,6 +1,0 @@
-import React from 'react'
-
-let RemoveTagIcon = props => (
-  <span className="tags-input-tag-remove fa fa-times" {...props} />
-)
-export default RemoveTagIcon
